src/test/javascript/spec/app/entities/categorie-tache/categorie-tache.service.spec.ts
import { TestBed, getTestBed } from '@angular/core/testing';
import { HttpClientTestingModule, HttpTestingController } from '@angular/common/http/testing';
import { CategorieTacheService } from 'app/entities/categorie-tache/categorie-tache.service';
import { ICategorieTache, CategorieTache } from 'app/shared/model/categorie-tache.model';
import { Couleur } from 'app/shared/model/enumerations/couleur.model';

describe('Service Tests', () => {
  describe('CategorieTache Service', () => {
    let injector: TestBed;
    let service: CategorieTacheService;
    let httpMock: HttpTestingController;
    let elemDefault: ICategorieTache;
    let expectedResult: ICategorieTache | ICategorieTache[] | boolean | null;

    beforeEach(() => {
      TestBed.configureTestingModule({
        imports: [HttpClientTestingModule]
      });
      expectedResult = null;
      injector = getTestBed();
      service = injector.get(CategorieTacheService);
      httpMock = injector.get(HttpTestingController);

      elemDefault = new CategorieTache(0, 'AAAAAAA', Couleur.ROUGE);
    });

    describe('Service methods', () => {
      it('should find an element', () => {
        const returnedFromService = Object.assign({}, elemDefault);

        service.find(123).subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'GET' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject(elemDefault);
      });

      it('should create a CategorieTache', () => {
        const returnedFromService = Object.assign(
          {
            id: 0
          },
          elemDefault
        );

        const expected = Object.assign({}, returnedFromService);

        service.create(new CategorieTache()).subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'POST' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject(expected);
      });

      it('should update a CategorieTache', () => {
        const returnedFromService = Object.assign(
          {
            libelle: 'BBBBBB',
            couleur: 'BBBBBB'
          },
          elemDefault
        );

        const expected = Object.assign({}, returnedFromService);

        service.update(expected).subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'PUT' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject(expected);
      });

      it('should return a list of CategorieTache', () => {
        const returnedFromService = Object.assign(
          {
            libelle: 'BBBBBB',
            couleur: 'BBBBBB'
          },
          elemDefault
        );

        const expected = Object.assign({}, returnedFromService);

        service.query().subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'GET' });
        req.flush([returnedFromService]);
        httpMock.verify();
        expect(expectedResult).toContainEqual(expected);
      });

      it('should delete a CategorieTache', () => {
        service.delete(123).subscribe(resp => (expectedResult = resp.ok));

        const req = httpMock.expectOne({ method: 'DELETE' });
        req.flush({ status: 200 });
        expect(expectedResult);
      });
    });

    afterEach(() => {
      httpMock.verify();
    });
  });
});
